security, ts, token and more updates

+++ server/controllers/categoriesController.js
@@ -1,4 +1,5 @@
 const categories = require('../models/categoriesModel')
+const { isValidId, isNonEmptyString } = require('../utils/validators')
 
 class categoriesController {
   async findAllCategories(req, res) {
@@ -26,6 +27,7 @@ class categoriesController {
 
   async deleteCategory(req, res) {
     let { name } = req.body  
+    if (!isNonEmptyString(name)) return res.status(400).send({ error: 'Invalid category name' })
 
     try {
       const removed = await categories.deleteOne({ name: name })
@@ -39,6 +41,7 @@ class categoriesController {
 
   async updateCategory(req, res) {
     let params = req.body 
+    if (!isValidId(params._id)) return res.status(400).send({ error: 'Invalid id' })
     
     try {
       const updated = await categories.updateOne(
@@ -53,7 +56,14 @@ class categoriesController {
   }
 
   async findOneCategory(req, res) {
-
+    try {
+      const cat = await categories.findOne({ name: req.params.category })
+      if (!cat) return res.status(404).send({ error: 'Category not found' })
+      res.send(cat)
+    }
+    catch (error) {
+      res.send({ error })
+    }
   }
 }
