+++ server/controllers/scoresController.js
@@ -1,4 +1,5 @@
 const scores = require('../models/scoresModel');
+const { isValidId } = require('../utils/validators');
 
 class scoresController {
   async findAllScores(req, res) {
@@ -24,6 +25,7 @@ class scoresController {
 
   async deleteScore(req, res) {
     let { _id } = req.body;
+    if (!isValidId(_id)) return res.status(400).send({ error: 'Invalid id' });
     try {
       const removed = await scores.deleteOne({ _id: _id });
       res.send({ removed });
@@ -35,6 +37,7 @@ class scoresController {
 
   async updateScore(req, res) {
     let params = req.body;  
+    if (!isValidId(params._id)) return res.status(400).send({ error: 'Invalid id' });
     
     try {
       const updated = await scores.updateOne(
